// server/enums/trainingAndQualificationsValue.ts
enum TrainingAndQualificationsValue {
  ADVANCED_EDUCATION = 'ADVANCED_EDUCATION',
  CSCS = 'CSCS',
  DRIVING_LICENSE = 'DRIVING_LICENSE',
  FIRST_AID = 'FIRST_AID',
  FOOD_HYGIENE = 'FOOD_HYGIENE',
  HEALTH_AND_SAFETY = 'HEALTH_AND_SAFETY',
  HGV_LICENSE = 'HGV_LICENSE',
  HIGHER_EDUCATION = 'HIGHER_EDUCATION',
  MACHINERY = 'MACHINERY',
  MANUAL = 'MANUAL',
  TRADE = 'TRADE',
  SCHOOL_EDUCATION = 'SCHOOL_EDUCATION',
  OTHER = 'OTHER',
  NONE = 'NONE',
}

export default TrainingAndQualificationsValue